fix folders structure, put dropbox on /paths
