+++ modules/paths/dropbox.py
@@ -20,3 +20,27 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with Hardcode-Tray. If not, see <http://www.gnu.org/licenses/>.
 """
+from os import path
+from sys import argv
+
+
+def get_dropbox_version(directory):
+    """
+    Correct the hardcoded dropbox directory.
+
+    Args:
+        directory(str): the default dropbox directory
+    """
+    version_file = directory.split("{dropbox_version}")[0].split("/")
+    del version_file[len(version_file) - 1]
+    version_file = "/".join(version_file) + "/VERSION"
+    if path.exists(version_file):
+        with open(version_file) as f:
+            return f.read()
+    return ""
+
+
+dropbox_path = argv[1]
+dropbox_path = dropbox_path .replace(
+    "{dropbox_version}", get_dropbox_version(dropbox_path))
+print(dropbox_path)
